+++ Code-DD-UFG/Back-End/sdd-ufg/src/main/java/br/ufg/inf/sdd_ufg/resource/KnowledgeGroupResource.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import javax.inject.Inject;
+import javax.persistence.EntityExistsException;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -18,7 +19,10 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import br.ufg.inf.sdd_ufg.dao.KnowledgeGroupDao;
+import br.ufg.inf.sdd_ufg.dao.TeacherDao;
 import br.ufg.inf.sdd_ufg.model.KnowledgeGroup;
+import br.ufg.inf.sdd_ufg.model.KnowledgeLevel;
+import br.ufg.inf.sdd_ufg.model.Teacher;
 import br.ufg.inf.sdd_ufg.resource.utils.ResultSetResponse;
 
 @Path("/knowledgeGroups")
@@ -26,18 +30,24 @@ import br.ufg.inf.sdd_ufg.resource.utils.ResultSetResponse;
 public class KnowledgeGroupResource extends AbstractResource {
 
 	private final KnowledgeGroupDao knowledgeGroupDao;
+	private final TeacherDao teacherDao;
 
     @Inject
-    public KnowledgeGroupResource(final KnowledgeGroupDao knowledgeGroupDao) {
+    public KnowledgeGroupResource(final KnowledgeGroupDao knowledgeGroupDao, final TeacherDao teacherDao) {
         this.knowledgeGroupDao = knowledgeGroupDao;
+        this.teacherDao = teacherDao;
     }
 	
     @GET
 	@Path("/{id}")
 	public Response retrieveKnowledgeGroupById(@PathParam("id") Long id, @Context final HttpServletRequest request) {
+    	if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+    	
 		KnowledgeGroup knowledgeGroup = knowledgeGroupDao.findById(id, 1);
 		if (knowledgeGroup == null) {
-			return Response.status(Response.Status.NOT_FOUND).build();
+			return getResourceNotFoundResponse();
 		}
 		return Response.ok(knowledgeGroup)
 				.build();
@@ -45,11 +55,11 @@ public class KnowledgeGroupResource extends AbstractResource {
     
 	@GET
     public Response retrieveAllKnowledgeGroups(@QueryParam("page") Integer page, @Context final HttpServletRequest request) {
-		List<KnowledgeGroup> knowledgeGroups = knowledgeGroupDao.findAll(0);
-		if (knowledgeGroups == null || knowledgeGroups.size() == 0) {
-			return Response.status(Response.Status.NOT_FOUND)
-					.build();
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
 		}
+		
+		List<KnowledgeGroup> knowledgeGroups = knowledgeGroupDao.findAll(0);
 		if (page == null) {
 			page = 1;
 		}
@@ -61,32 +71,57 @@ public class KnowledgeGroupResource extends AbstractResource {
 	
 	@POST
 	public Response insertKnowledgeGroup(@Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		KnowledgeGroup knowledgeGroup;
 		try {
 			knowledgeGroup = retrieveKnowledgeGroupFromJson(request);
 			knowledgeGroupDao.insert(knowledgeGroup);
+			
+			updateTeachersKnowledge(knowledgeGroup);
+		} catch (EntityExistsException eee) {
+			return getInsertErrorResponse();
 		} catch (Exception e) {
-			return Response.status(Response.Status.BAD_REQUEST)
-					.build();
+			return getBadRequestResponse();
 		}
 		return Response.ok(knowledgeGroup)
 				.build();
 	}
 	
+	private void updateTeachersKnowledge(KnowledgeGroup kg) {
+		List<Teacher> teachers = teacherDao.findAll(2);
+		for (Teacher teacher : teachers) {
+			KnowledgeLevel kl = new KnowledgeLevel();
+			kl.setLevel(3);
+			kl.setTeacher(teacher);
+			kl.setKnowledgeGroup(kg);
+			
+			teacher.getKnowledgeLevels().add(kl);
+			teacherDao.update(teacher);
+		}
+		
+	}
+	
 	@PUT
 	@Path("/{id}")
 	public Response updateKnowledgeGroup(@PathParam("id") Long id, @Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		KnowledgeGroup knowledgeGroup;
 		try {
 			knowledgeGroup = retrieveKnowledgeGroupFromJson(request);
 			knowledgeGroup.setId(id);
 			knowledgeGroupDao.update(knowledgeGroup);
+		} catch (EntityExistsException eee) {
+			return getInsertErrorResponse();
 		} catch (IllegalArgumentException iae) {
-			return Response.status(Response.Status.NOT_FOUND)
-					.build();
+			return getResourceNotFoundResponse();
 		} catch (Exception e) {
-			return Response.status(Response.Status.BAD_REQUEST)
-					.build();
+			return getBadRequestResponse();
 		}
 		return Response.ok(knowledgeGroup)
 				.build();
@@ -97,7 +132,6 @@ public class KnowledgeGroupResource extends AbstractResource {
 		
 		KnowledgeGroup knowledgeGroup = new KnowledgeGroup();
 		knowledgeGroup.setName(content.get("name").toString());
-		//knowledgeGroup.setGrade(content.get("cpf").toString());
 		
 		return knowledgeGroup;
 	}
@@ -105,11 +139,14 @@ public class KnowledgeGroupResource extends AbstractResource {
 	@DELETE
 	@Path("/{id}")
 	public Response deleteKnowledgeGroup(@PathParam("id") Long id, @Context final HttpServletRequest request) {
+		if (validateSession(request)) {
+			return getAuthenticationErrorResponse();
+		}
+		
 		try {
 			knowledgeGroupDao.delete(id);
 		} catch (IllegalArgumentException iae) {
-			return Response.status(Response.Status.NOT_FOUND)
-					.build();
+			return getResourceNotFoundResponse();
 		}
 		return Response.status(Response.Status.NO_CONTENT)
 				.build();
